requirements.txt
Flask==3.0.2
fpdf2==2.7.7
Pillow==10.2.0
Werkzeug==3.0.1